requirements.txt
python-socketio[client]==5.11.4
pyobjc-framework-Quartz==10.1
